connect-guernsey-complete/connect-guernsey/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
supabase==2.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic[email]==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
python-slugify==8.0.4
